requirements.txt
tornado==6.3.1
mysql-connector-python==8.0.33
psutil==5.9.5
